Random priority if same arrival time

+++ Main.py
@@ -50,20 +50,20 @@ def display_universe(universe):
     """
 
 # number of games = number of univers
-number_of_univers = 2
+number_of_univers = 1
 
 # if flag_fair is True, each univers will be played with players in each starting position
 # Warning : this option multiply the number of game (= number_of_univers) depending of the number of player
 #           2 players => x2 ; 3 players => x6 ; 4 players => x24
-flag_fair = True
+flag_fair = False
 
 # -------------------------------------------------------------------------------------------------
 # creation of the neutral player... which is not a player
 player_neutral = Player(ai=None, name="neutral", color=COLOR_PLAYER_LIGHT[0])
 # creation of the players
 players = [
-    Player(ai=AI_dumb, name="AI_1", color=COLOR_PLAYER_LIGHT[1]),
-    Player(ai=AI_dumb, name="AI_2", color=COLOR_PLAYER_LIGHT[2]),
+    # Player(ai=AI_dumb, name="AI_1", color=COLOR_PLAYER_LIGHT[1]),
+    # Player(ai=AI_dumb, name="AI_2", color=COLOR_PLAYER_LIGHT[2]),
     Player(ai=AI_dumb, name="AI_3", color=COLOR_PLAYER_LIGHT[3]),
     Player(ai=AI_dumb, name="AI_4", color=COLOR_PLAYER_LIGHT[4])
 ]
@@ -79,7 +79,7 @@ for game_number in range(number_of_univers):
             universe = Universe()
             universe.big_bang(
                 size=10,
-                nb_planets=10,
+                nb_planets=5,
                 size_planet_max=3,
                 list_players=list_players,
                 player_neutral=player_neutral,
@@ -114,5 +114,5 @@ for game_number in range(number_of_univers):
             break
 
 # history
-if number_of_univers == 1:
+if number_of_univers == 1 and not flag_fair:  # i.e. only one simulation
     pickle.dump(timeline, open(HISTORY_FILE, 'wb'))
